Move tests from `tracker/mod.rs` into `tracker/tests.rs`
